+++ src/app/components/main-layout/header/ribbon.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { BasicsProps } from '@/app/components/ui/types';
-import { cn } from '@/utils';
+import { cn } from '@/lib/utils';
 
 type IHeaderRibbonProps = BasicsProps;
