+++ src/app/components/main-layout/profile/wrapper.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 
 import { BasicsProps } from '@/app/components/ui/types';
-import { cn } from '@/utils';
+import { cn } from '@/lib/utils';
 
 type IProfileWrapperProps = BasicsProps<'div', 'children'>;
