+++ src/app/components/ui/hover-card.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 
 import * as HoverCardPrimitive from '@radix-ui/react-hover-card';
 
-import { cn } from '@/utils';
+import { cn } from '@/lib/utils';
 
 const HoverCard = HoverCardPrimitive.Root;
